# apps/api/openapi/khana.v1.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/tenant": {
      "get": {
        "description": "Returns public tenant context used by auth and onboarding flows before a JWT exists.",
        "operationId": "Auth_getTenantContext",
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "description": "Optional tenant hint for public auth and tenant-resolution flows when JWT context is not available.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tenant context resolved from request hints.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/AuthTenantContextDoc"
                    }
                  ],
                  "example": {
                    "id": "11111111-1111-4111-8111-111111111111",
                    "name": "Khana Padel Club",
                    "slug": "khana-padel-club",
                    "timezone": "Asia/Riyadh"
                  }
                }
              }
            }
          }
        },
        "security": [],
        "summary": "Resolve tenant context from the current request",
        "tags": ["Auth"]
      }
    },
    "/api/v1/auth/tenant/resolve": {
      "get": {
        "operationId": "Auth_resolveTenantBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "query",
            "description": "Public tenant slug used to resolve auth context.",
            "schema": {
              "example": "khana-padel-club",
              "type": "string"
            }
          },
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tenant context for the supplied public slug.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/AuthTenantContextDoc"
                    }
                  ],
                  "example": {
                    "id": "11111111-1111-4111-8111-111111111111",
                    "name": "Khana Padel Club",
                    "slug": "khana-padel-club",
                    "timezone": "Asia/Riyadh"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed or the payload is invalid for this operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwaggerErrorResponse"
                },
                "example": {
                  "statusCode": 400,
                  "message": ["email must be an email"],
                  "error": "Bad Request",
                  "timestamp": "2026-03-11T12:00:00.000Z",
                  "path": "/api/v1/auth/login"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests were made and throttling was applied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwaggerErrorResponse"
                },
                "example": {
                  "statusCode": 429,
                  "message": "Too many requests",
                  "error": "Too Many Requests",
                  "timestamp": "2026-03-11T12:00:00.000Z",
                  "path": "/api/v1/auth/login"
                }
              }
            }
          }
        },
        "security": [],
        "summary": "Resolve a tenant by public slug",
        "tags": ["Auth"]
      }
    },
    "/api/v1/auth/signup-owner": {
      "post": {
        "operationId": "Auth_signupOwner",
        "parameters": [
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Owner signup payload with the initial tenant context.",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/SignupOwnerDto"
                  }
                ],
                "example": {
                  "workspaceName": "Khana Padel Club",
                  "workspaceSlug": "khana-padel-club",
                  "name": "Malek Elkhazendar",
                  "email": "owner@example.test",
                  "phone": "+966500000000",
                  "password": "example-password-not-real"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Owner account created and initial auth tokens returned.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/AuthLoginResponseDoc"
                    }
                  ],
                  "example": {
                    "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.access",
                    "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.refresh",
                    "expiresIn": 900,
                    "user": {
                      "id": "22222222-2222-4222-8222-222222222222",
                      "tenantId": "11111111-1111-4111-8111-111111111111",
                      "email": "owner@example.test",
                      "name": "Malek Elkhazendar",
                      "phone": "+966500000000",
                      "role": "OWNER",
                      "isActive": true,
                      "onboardingCompleted": true,
                      "lastLoginAt": "2030-06-15T18:00:00.000Z",
                      "createdAt": "2030-06-01T10:00:00.000Z",
                      "updatedAt": "2030-06-15T18:00:00.000Z"
                    },
                    "tenant": {
                      "id": "11111111-1111-4111-8111-111111111111",
                      "name": "Khana Padel Club",
                      "slug": "khana-padel-club",
                      "timezone": "Asia/Riyadh"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed or the payload is invalid for this operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwaggerErrorResponse"
                },
                "example": {
                  "statusCode": 400,
                  "message": ["email must be an email"],
                  "error": "Bad Request",
                  "timestamp": "2026-03-11T12:00:00.000Z",
                  "path": "/api/v1/auth/login"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with current resource state or business rules.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwaggerErrorResponse"
                },
                "example": {
                  "statusCode": 409,
                  "message": "Booking conflicts with an existing reservation",
                  "error": "Conflict",
                  "timestamp": "2026-03-11T12:00:00.000Z",
                  "path": "/api/v1/bookings"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests were made and throttling was applied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwaggerErrorResponse"
                },
                "example": {
                  "statusCode": 429,
                  "message": "Too many requests",
                  "error": "Too Many Requests",
                  "timestamp": "2026-03-11T12:00:00.000Z",
                  "path": "/api/v1/auth/login"
                }
              }
            }
          }
        },
        "security": [],
        "summary": "Create a new owner account and tenant",
        "tags": ["Auth"]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "Auth_register",
        "parameters": [
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "description": "Optional tenant hint for public auth and tenant-resolution flows when JWT context is not available.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "User registration payload for an existing tenant.",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/RegisterDto"
                  }
                ],
                "example": {
                  "email": "manager@example.test",
                  "password": "example-password-not-real",
                  "name": "Abeer Alotaibi",
                  "phone": "+966500000000"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered and auth tokens returned.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/AuthLoginResponseDoc"
                    }
                  ],
                  "example": {
                    "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.access",
                    "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.refresh",
                    "expiresIn": 900,
                    "user": {
                      "id": "22222222-2222-4222-8222-222222222222",
                      "tenantId": "11111111-1111-4111-8111-111111111111",
                      "email": "owner@example.test",
                      "name": "Malek Elkhazendar",
                      "phone": "+966500000000",
                      "role": "OWNER",
                      "isActive": true,
                      "onboardingCompleted": true,
                      "lastLoginAt": "2030-06-15T18:00:00.000Z",
                      "createdAt": "2030-06-01T10:00:00.000Z",
                      "updatedAt": "2030-06-15T18:00:00.000Z"
                    },
                    "tenant": {
                      "id": "11111111-1111-4111-8111-111111111111",
                      "name": "Khana Padel Club",
                      "slug": "khana-padel-club",
                      "timezone": "Asia/Riyadh"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed or the payload is invalid for this operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwaggerErrorResponse"
                },
                "example": {
                  "statusCode": 400,
                  "message": ["email must be an email"],
                  "error": "Bad Request",
                  "timestamp": "2026-03-11T12:00:00.000Z",
                  "path": "/api/v1/auth/login"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with current resource state or business rules.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwaggerErrorResponse"
                },
                "example": {
                  "statusCode": 409,
                  "message": "Booking conflicts with an existing reservation",
                  "error": "Conflict",
                  "timestamp": "2026-03-11T12:00:00.000Z",
                  "path": "/api/v1/bookings"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests were made and throttling was applied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwaggerErrorResponse"
                },
                "example": {
                  "statusCode": 429,
                  "message": "Too many requests",
                  "error": "Too Many Requests",
                  "timestamp": "2026-03-11T12:00:00.000Z",
                  "path": "/api/v1/auth/login"
                }
              }
            }
          }
        },
        "security": [],
        "summary": "Register a user for an existing tenant",
        "tags": ["Auth"]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "Auth_login",
        "parameters": [
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "description": "Optional tenant hint for public auth and tenant-resolution flows when JWT context is not available.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Email/password login payload with an optional public tenant hint.",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/LoginDto"
                  }
                ],
                "example": {
                  "email": "owner@example.test",
                  "password": "example-password-not-real",
                  "subdomain": "khana-padel-club"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Access token, refresh token, and current user context.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/AuthLoginResponseDoc"
                    }
                  ],
                  "example": {
                    "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.access",
                    "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.refresh",
                    "expiresIn": 900,
                    "user": {
                      "id": "22222222-2222-4222-8222-222222222222",
                      "tenantId": "11111111-1111-4111-8111-111111111111",
                      "email": "owner@example.test",
                      "name": "Malek Elkhazendar",
                      "phone": "+966500000000",
                      "role": "OWNER",
                      "isActive": true,
                      "onboardingCompleted": true,
                      "lastLoginAt": "2030-06-15T18:00:00.000Z",
                      "createdAt": "2030-06-01T10:00:00.000Z",
                      "updatedAt": "2030-06-15T18:00:00.000Z"
                    },
                    "tenant": {
                      "id": "11111111-1111-4111-8111-111111111111",
                      "name": "Khana Padel Club",
                      "slug": "khana-padel-club",
                      "timezone": "Asia/Riyadh"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed or the payload is invalid for this operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwaggerErrorResponse"
                },
                "example": {
                  "statusCode": 400,
                  "message": ["email must be an email"],
                  "error": "Bad Request",
                  "timestamp": "2026-03-11T12:00:00.000Z",
                  "path": "/api/v1/auth/login"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required or the supplied token is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwaggerErrorResponse"
                },
                "example": {
                  "statusCode": 401,
                  "message": "Invalid email or password",
                  "error": "Unauthorized",
                  "timestamp": "2026-03-11T12:00:00.000Z",
                  "path": "/api/v1/auth/login"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests were made and throttling was applied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwaggerErrorResponse"
                },
                "example": {
                  "statusCode": 429,
                  "message": "Too many requests",
                  "error": "Too Many Requests",
                  "timestamp": "2026-03-11T12:00:00.000Z",
                  "path": "/api/v1/auth/login"
                }
              }
            }
          }
        },
        "security": [],
        "summary": "Authenticate with email and password",
        "tags": ["Auth"]
      }
    },
...
    "/api/v1/auth/logout-all-devices": {
      "post": {
        "operationId": "Auth_logoutAllDevices",
        "parameters": [],
        "responses": {
          "204": {
            "description": "All other device sessions were revoked."
          },
          "401": {
            "description": "Authentication is required or the supplied token is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwaggerErrorResponse"
                },
                "example": {
                  "statusCode": 401,
                  "message": "Invalid email or password",
                  "error": "Unauthorized",
                  "timestamp": "2026-03-11T12:00:00.000Z",
                  "path": "/api/v1/auth/login"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt-bearer": []
          }
        ],
        "summary": "Log out all devices except the current session",
        "tags": ["Auth"]
      }
    },
...
  },
  "info": {
    "title": "Khana API",
    "description": "Internal OpenAPI documentation for validating and testing Khana API endpoints before frontend integration.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Relative server root for exported specs and local docs"
    }
  ],
  "components": {
    "securitySchemes": {
      "jwt-bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Use Authorization: Bearer <token> for protected endpoints."
      }
    },
    "schemas": {
      "AuthTenantContextDoc": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Khana Padel Club"
          },
          "slug": {
            "type": "string",
            "example": "khana-padel-club"
          },
          "timezone": {
            "type": "string",
            "example": "Asia/Riyadh"
          }
        },
        "required": ["id", "name", "slug", "timezone"]
      },
      "SwaggerErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "oneOf": [
              {
                "type": "string",
                "example": "Validation failed"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": ["email must be an email"]
              }
            ]
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-03-11T12:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/bookings"
          }
        },
        "required": ["statusCode", "message", "error", "timestamp", "path"]
      },
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email used for authentication.",
            "example": "owner@example.test"
          },
          "password": {
            "type": "string",
            "description": "Plain-text password submitted for login.",
            "minLength": 8,
            "example": "example-password-not-real"
          },
          "subdomain": {
            "type": "string",
            "description": "Optional tenant subdomain hint for public login flows.",
            "example": "riyadh-club"
          }
        },
        "required": ["email", "password"]
      },
...
    }
  },
  "security": [
    {
      "jwt-bearer": []
    }
  ]
}